Fix bug (due to incorrect usage of array indexing vs pointer arithmetic) and get tokenizer working.

+++ src/tokenize.h
@@ -25,8 +25,8 @@ typedef struct
 typedef struct
 {
     Token *tokens;
-    size_t size;
-    size_t cap;
+    unsigned int size;
+    unsigned int cap;
 } TokenArray;
 
 TokenArray *tokenize(char input[]);
